requirements.txt
aiobotocore==2.9.0
aiofiles==23.2.1
aiohttp==3.9.1
aioitertools==0.11.0
aiosignal==1.3.1
alabaster==0.7.13
anyio==3.7.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
astroid==2.15.5
asttokens==2.4.1
async-lru==2.0.4
attrs==23.1.0
autopep8==2.0.2
Babel==2.14.0
beautifulsoup4==4.12.2
bidict==0.22.1
bleach==6.1.0
botocore==1.33.13
build==1.0.3
cellpose==2.2.3
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==2.2.1
colorama==0.4.6
comm==0.2.0
contourpy==1.2.0
cycler==0.12.1
dask==2023.12.0
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.6
dnspython==2.4.2
docstring-inheritance==2.1.1
docstring-to-markdown==0.12
docutils==0.20.1
ecdsa==0.18.0
email-validator==2.1.0.post1
executing==2.0.1
fastapi==0.103.0
fastjsonschema==2.19.0
fastremap==1.14.0
filelock==3.13.1
flake8==6.0.0
fonttools==4.46.0
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2023.12.2
ftd2xx==1.3.8
h11==0.14.0
h5py==3.11.0
idna==3.6
imagecodecs==2023.9.18
imageio==2.31.1
imagesize==1.4.1
imglyb==2.1.0
imjoy==0.11.20
imjoy-jupyter-extension==0.3.0
imjoy-rpc==0.5.16
importlib-metadata==7.0.0
ipykernel==6.26.0
ipython==8.17.2
ipython-genutils==0.2.0
isoduration==20.11.0
isort==5.12.0
jedi==0.18.2
jgo==1.0.5
Jinja2==3.1.2
jmespath==1.0.1
JPype1==1.4.1
json5==0.9.14
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
jupyter-events==0.6.3
jupyter-lsp==2.2.1
jupyter_client==8.6.0
jupyter_core==5.5.0
jupyter_server==2.6.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.9
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
jupyterlab_server==2.25.2
kiwisolver==1.4.5
labeling==0.1.13
lazy-object-proxy==1.9.0
lazy_loader==0.3
llvmlite==0.41.1
locket==1.0.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.1
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
mistune==3.0.2
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.4
mypy-extensions==1.0.0
natsort==8.4.0
nbclient==0.9.0
nbconvert==7.12.0
nbformat==5.9.2
ndtiff==2.2.1
nest-asyncio==1.5.8
networkx==3.2.1
notebook==7.0.6
notebook_shim==0.2.3
np-image-buffer==0.1.1
npe2==0.7.0
numba==0.58.1
numpy==1.26.2
numpydoc==1.5.0
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78
overrides==7.4.0
packaging==23.2
pandas==2.0.2
pandocfilters==1.5.0
parso==0.8.3
partd==1.4.1
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.8.0
pluggy==1.2.0
prometheus-client==0.19.0
prompt-toolkit==3.0.40
psutil==5.9.6
psygnal==0.9.5
pure-eval==0.2.2
pyasn1==0.5.1
pycodestyle==2.10.0
pycparser==2.21
pycromanager==0.29.9
pydantic==1.10.13
pydocstyle==6.3.0
pyFirmata==1.1.0
pyflakes==3.0.1
Pygments==2.16.1
pyimagej==1.4.1
pylint==2.17.4
pymmcore==11.1.1.71.0
pyparsing==3.1.1
pyproject_hooks==1.0.0
pyqtgraph==0.13.7
pyserial==3.5
PySide6==6.5.1.1
PySide6-Addons==6.5.1.1
PySide6-Essentials==6.5.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.0.0
python-jose==3.3.0
python-json-logger==2.0.7
python-lsp-jsonrpc==1.0.0
python-lsp-server==1.7.3
python-socketio==5.0.4
pytomlpp==1.0.13
pytoolconfig==1.2.5
pytz==2023.3.post1
PyWavelets==1.5.0
pywin32==306
pywinpty==2.0.12
PyYAML==6.0.1
pyzmq==25.1.2
qtconsole==5.4.3
QtPy==2.3.1
referencing==0.32.0
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.0
roifile==2023.8.30
rope==1.9.0
rpds-py==0.13.2
rsa==4.9
scikit-image==0.21.0
scipy==1.10.1
screeninfo==0.8.1
scyjava==1.9.1
Send2Trash==1.8.2
shiboken6==6.5.1.1
shortuuid==1.0.11
six==1.16.0
sniffio==1.3.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.5
Sphinx==7.2.6
sphinxcontrib-applehelp==1.0.7
sphinxcontrib-devhelp==1.0.5
sphinxcontrib-htmlhelp==2.0.4
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.6
sphinxcontrib-serializinghtml==1.1.9
stack-data==0.6.3
starlette==0.27.0
sympy==1.12
terminado==0.18.0
tifffile==2023.4.12
tinycss2==1.2.1
tomli==2.0.1
tomlkit==0.11.8
toolz==0.12.0
torch==2.1.1
tornado==6.3.3
tqdm==4.66.1
traitlets==5.13.0
typer==0.9.0
types-python-dateutil==2.8.19.14
typing_extensions==4.9.0
tzdata==2023.3
ujson==5.8.0
uri-template==1.3.0
urllib3==2.0.7
uvicorn==0.24.0.post1
wcwidth==0.2.9
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
websockets==12.0
whatthepatch==1.0.5
wrapt==1.15.0
xarray==2023.12.0
yapf==0.40.1
yarl==1.9.4
zipp==3.17.0
